% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_rss.R
\name{get_default_feeds}
\alias{get_default_feeds}
\title{Get default security RSS feeds}
\usage{
get_default_feeds()
}
\value{
Named character vector of default RSS feed URLs.
}
\description{
Get default security RSS feeds
}
