% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_github.R
\name{get_default_github_queries}
\alias{get_default_github_queries}
\title{Get default GitHub search queries}
\usage{
get_default_github_queries()
}
\value{
Character vector of default GitHub search queries.
}
\description{
Returns a curated set of search phrases for offensive security tools.
}
