% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_github.R
\name{get_default_github_seed_names}
\alias{get_default_github_seed_names}
\title{Get default GitHub seed tool names}
\usage{
get_default_github_seed_names()
}
\value{
Character vector of default seed tool names.
}
\description{
Returns a curated list of well-known offensive security tool names that can
be used as additional GitHub search seeds.
}
