% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_packetstorm.R
\name{get_default_packetstorm_categories}
\alias{get_default_packetstorm_categories}
\title{Get default PacketStorm categories}
\usage{
get_default_packetstorm_categories()
}
\value{
Character vector of default PacketStorm categories.
}
\description{
Get default PacketStorm categories
}
