% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_mitre.R
\name{get_mitre_matrix}
\alias{get_mitre_matrix}
\title{Build a MITRE ATT&CK matrix table}
\usage{
get_mitre_matrix(stix_data)
}
\arguments{
\item{stix_data}{Parsed STIX bundle.}
}
\value{
Tibble joining techniques to tactics.
}
\description{
Build a MITRE ATT&CK matrix table
}
