% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_mitre.R
\name{parse_mitre_tactics}
\alias{parse_mitre_tactics}
\title{Parse MITRE ATT&CK tactics}
\usage{
parse_mitre_tactics(stix_data)
}
\arguments{
\item{stix_data}{Parsed STIX bundle.}
}
\value{
Tibble with normalized MITRE tactics.
}
\description{
Parse MITRE ATT&CK tactics
}
