% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etl_mitre.R
\name{parse_mitre_techniques}
\alias{parse_mitre_techniques}
\title{Parse MITRE ATT&CK techniques and sub-techniques}
\usage{
parse_mitre_techniques(stix_data)
}
\arguments{
\item{stix_data}{Parsed STIX bundle.}
}
\value{
Tibble with normalized techniques in long format by tactic.
}
\description{
Parse MITRE ATT&CK techniques and sub-techniques
}
