% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_core.R
\name{require_env_var}
\alias{require_env_var}
\title{Read a required environment variable}
\usage{
require_env_var(name)
}
\arguments{
\item{name}{Environment variable name.}
}
\value{
Character scalar with the environment variable value.
}
\description{
Read a required environment variable
}
