% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OffensiveToolMapper-package.R
\docType{package}
\name{OffensiveToolMapper-package}
\alias{OffensiveToolMapper}
\alias{OffensiveToolMapper-package}
\title{OffensiveToolMapper}
\description{
Bootstrap package for collecting, validating, enriching, and mapping
offensive cybersecurity tools to MITRE ATT&CK.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://KNikitaaa.github.io/CyberSecML-NetAdmins/}
  \item \url{https://github.com/KNikitaaa/CyberSecML-NetAdmins}
}

}
\author{
\strong{Maintainer}: KNikitaaa Maintainer \email{KNikitaaa@users.noreply.github.com}

}
\keyword{internal}
