% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OffensiveToolMapper-package.R
\docType{package}
\name{OffensiveToolMapper-package}
\alias{OffensiveToolMapper}
\alias{OffensiveToolMapper-package}
\title{OffensiveToolMapper}
\description{
Bootstrap package for collecting, validating, enriching, and mapping
offensive cybersecurity tools to MITRE ATT&CK.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://w8ofdope.github.io/OffensiveToolMapper/}
  \item \url{https://github.com/w8ofdope/OffensiveToolMapper}
}

}
\author{
\strong{Maintainer}: w8ofdope Maintainer \email{69384386+w8ofdope@users.noreply.github.com}

}
\keyword{internal}
